refactor: extract pure view composable
